% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_pars.R
\name{add_prm_association}
\alias{add_prm_association}
\title{Describe parameter associations}
\usage{
add_prm_association(
  xpdb_x,
  ...,
  .problem = NULL,
  .subprob = NULL,
  .method = NULL,
  quiet
)
}
\arguments{
\item{xpdb_x}{h}

\item{...}{h}

\item{.problem}{h}

\item{.subprob}{h}

\item{.method}{h}

\item{quiet}{h}
}
\description{
The relationship between structural parameters and
omega parameters can be described. This is useful if
it deviates from the typical log-normal.

Default transformations are those that are built into
\code{pmxcv}, but see examples for how associations can be
described for other relationships.
}
\details{
At time of writing, the built-in distributions for \code{pmxcv} are below.
Those marked with an asterisk require a fixed effect parameter to calculate CV.
\itemize{
\item \code{log} typical log-normal. Optional \code{exact} parameter.
\item \code{logexp}* modified log-normal \code{log(1+X)}
\item \code{logit}* logit-normal
\item \code{arcsin}* arcsine-transform
\item \code{nmboxcox}* Box-Cox transform as typically implemented in pharmacometrics.
Requires a \code{lambda} parameter.
}

To pass a custom parameter, use \code{custom} transform, and pass \code{pdist} and \code{qdist} to
that transform. See Examples.
}
\examples{
# example codes


}
\references{
Prybylski, J.P. Reporting Coefficient of Variation for Logit, Box-Cox and
Other Non-log-normal Parameters. Clin Pharmacokinet 63, 133–135 (2024).
https://doi.org/10.1007/s40262-023-01343-2
}
\seealso{
\code{\link[pmxcv:dist.intcv]{pmxcv::dist.intcv()}}
}
