% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{as.xpdb}
\alias{as.xpdb}
\title{Convert an object to an \code{xpose_data} and \code{xp_xtras} object}
\usage{
as.xpdb(x)
}
\arguments{
\item{x}{<\code{list}> object with \code{xpose_data} and \code{xp_xtras} components}
}
\value{
<\code{\link[xpose:xpose_data]{xpose_data}}> and <\code{xp_xtras}> object
}
\description{
This function masks the default in \code{xpose} package, adding the
\code{xp_xtras} class to default \code{xpose_data} objects.
}
\examples{

c()
}
