% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{backfill_iofv}
\alias{backfill_iofv}
\title{Add individual objective function to data}
\usage{
backfill_iofv(xpdb, .problem = NULL, .subprob = NULL, .label = "iOFV")
}
\arguments{
\item{xpdb}{<\code{xpose_data}> or <\code{xp_xtras}> object}

\item{.label}{The name of the new column. \code{iOFV} is the default.}
}
\value{
<\code{xp_xtras}> object with new column in the data and a
column with \code{iofv} var type.
}
\description{
Add individual objective function to data
}
\details{
This function will only work for objects with software listed as
\code{nonmem}, which has a \code{phi} file and with an \code{OBJ} column in that
file.
}
\examples{

xpdb_x \%>\%
  backfill_iofv() \%>\%
  list_vars()

}
