% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2.R
\name{backfill_nlmixr2_props}
\alias{backfill_nlmixr2_props}
\title{Populate some properties from nlmixr2 fit}
\usage{
backfill_nlmixr2_props(xpdb)
}
\arguments{
\item{xpdb}{<\code{xpose_data}> object}
}
\description{
Populate some properties from nlmixr2 fit
}
\details{
This function will currently backfill:
\itemize{
\item condn
\item nsig
}
}
\examples{

xpdb_nlmixr2 \%>\%
  set_prop(condn = "not implemented") \%>\%
  get_prop("condn")

xpdb_nlmixr2 \%>\%
  set_prop(condn = "not implemented") \%>\%
  backfill_nlmixr2_props() \%>\%
  get_prop("condn")

}
