% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confmatr_by_threshold}
\alias{confmatr_by_threshold}
\title{Calculate all binary confusion matrix indices}
\usage{
confmatr_by_threshold(
  test_vec,
  true_vec,
  ...,
  threshold = 0.5,
  pos_val = 1,
  prepend = "",
  cols = dplyr::everything()
)
}
\arguments{
\item{test_vec}{<\code{double}> Vector of probabilities}

\item{true_vec}{<\code{integer}> Vector true values}

\item{...}{For future extension}

\item{threshold}{<\code{double}> Number that defines probability as positive.}

\item{pos_val}{<\code{integer}> Positive value in \code{true_vec}}

\item{prepend}{<\code{character}> Preprend column names with this to prevent name conflicts}
}
\value{
dataframe of confusion matrix indices
}
\description{
Calculate all binary confusion matrix indices
}
\keyword{internal}
