% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_formula_list}
\alias{is_formula_list}
\title{Convenience functions used in package}
\usage{
is_formula_list(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Convenience functions used in package
}
