% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covars.R
\name{covars}
\alias{covars}
\alias{set_var_types_x}
\title{Set variable types}
\usage{
set_var_types_x(xpdb, .problem = NULL, ..., auto_factor = TRUE, quiet)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object.}

\item{.problem}{The problem number to which the edits will be applied.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Passed to \code{\link[xpose]{set_var_types}}.}

\item{auto_factor}{With \code{set_var_types} only. If \code{TRUE} new columns assigned to the type 'catcov' will be converted to
factor.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
An xpose_data object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[xpose]{set_var_types}} wrapper that accepts tidyselect syntax.
}
