% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xset_diagram.R
\name{diagram_lineage}
\alias{diagram_lineage}
\title{Visualize \code{xpose_set}}
\usage{
diagram_lineage(xpdb_s, ...)
}
\arguments{
\item{xpdb_s}{<\code{xpose_set}> object}

\item{...}{For later expansion. Will be ignored.}
}
\value{
A DiagrammeR-compliant graph object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

In its current state, this function is intended to
provide a simple visual representation of an \code{xpose_set}.
Functionality and aesthetic enhancements are expected in
future releases.
}
\examples{

diagram_lineage(pheno_set) \%>\%
  DiagrammeR::render_graph(layout="tree")

}
