% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xset_features.R
\name{diff.xpose_set}
\alias{diff.xpose_set}
\title{Display deltaOFV values across \code{xpose_set}}
\usage{
\method{diff}{xpose_set}(x, ...)
}
\arguments{
\item{x}{<\code{xpose_set}> object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Passed to <\code{\link{xset_lineage}}>.
\code{.spinner=FALSE} can also be set here.}
}
\value{
<\code{numeric}> vector of deltaOFV values
}
\description{
If no base model is provided, and if lineage is unclear,
the first model in the \code{xpose_set} is used as the base model.
}
\examples{

c()
}
