% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{duplicated.xpose_set}
\alias{duplicated.xpose_set}
\title{Check if any xpose_data objects are repeated in xpose_set}
\usage{
\method{duplicated}{xpose_set}(xpdb_s, ...)
}
\description{
Check if any xpose_data objects are repeated in xpose_set
}
