% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_set.R
\name{expose_property}
\alias{expose_property}
\title{Expose a property of xpdb objects in an xpose_set}
\usage{
expose_property(xpdb_s, ..., .get_fn = NULL)
}
\arguments{
\item{xpdb_s}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more properties to expose, or passed to \verb{get_*()} function.}

\item{.get_fn}{<\code{\link{function}}> Function in the get_*() family (with a single return value) to use for extraction.}
}
\value{
An \code{xpose_set} object with the properties exposed
}
\description{
Expose a property of xpdb objects in an xpose_set
}
\details{
The property returned will be top-level, and to avoid conflicting names will be preprended by an underscore.
}
\examples{
data("xpdb_ex_pk", package = "xpose")
set <- xpose_set(xpdb_ex_pk, xpdb_ex_pk2 = xpdb_ex_pk)

set <- expose_property(set, descr)
set$xpdb_ex_pk$_descr

set <- expose_property(set, .get_fun = get_shk)
set$xpdb_ex_pk$_etashk

set <- expose_property(set, wh="eps" .get_fun = get_shk)
set$xpdb_ex_pk$_epsshk

}
