% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_set.R
\name{filter.xpose_set}
\alias{filter.xpose_set}
\title{Filtration method for xpose_set}
\usage{
\method{filter}{xpose_set}(xpdb_s, ..., .rowwise = FALSE)
}
\arguments{
\item{xpdb_s}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> (passed through to <\code{\link[dplyr:filter]{dplyr::filter}}>)}

\item{.rowwise}{<\code{\link{logical}}> Should the mutation be applied rowwise? (default: \code{FALSE})}
}
\description{
Filtration method for xpose_set
}
\examples{
xpdb_set \%>\%
  filter(label=="mod1")

xpdb_set \%>\%
  filter(length(parent)>1, .rowwise=TRUE)


}
