% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{focus_xpdb}
\alias{focus_xpdb}
\alias{unfocus_xpdb}
\alias{focused_xpdbs}
\alias{focus_function}
\alias{focus_qapply}
\title{Focus on an xpdb object in an xpose_set}
\usage{
focus_xpdb(xpdb_s, ..., .add = FALSE)

unfocus_xpdb(xpdb_s)

focused_xpdbs(xpdb_s)

focus_function(xpdb_s, fn, ...)

focus_qapply(xpdb_s, fn, ..., .mods = everything())
}
\arguments{
\item{xpdb_s}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more xpdb objects to focus on}

\item{.add}{<\code{\link{logical}}> Should the focus be added to the existing focus? (default: \code{FALSE})}

\item{fn}{<\code{function}> to apply to focused \code{xpose_data} objects}

\item{.mods}{<\code{tidyselect}> Model names in set to quick-apply a function. See Details.}
}
\value{
An \code{xpose_set} object with the focused xpdb object(s) transformed in place, or,
for functions that do not return an \code{xpose_data}/\code{xp_xtras} object, the output of \code{fn}
(or a named list of outputs, if multiple elements are focused)
}
\description{
For piping, set is passed, but with S3 method transformations are applied to the focused \code{xpdb} object.
}
\details{
While these functions are used internally, it is recognized that they may have value
in user scripting. It is hoped these are self-explanatory, but the examples should
address common uses.

\emph{Note:} \code{focus_qapply()} (re)focuses as specified in \code{.mods} and then un-focuses all elements
of the set so should only be used in the case where a quick application
suffices. Otherwise, focusing with a sequence of \code{focus_function} calls
(or a monolithic single \code{focus_function} call with a custom function)
should be preferred.

\code{focus_function()}/\code{focus_qapply()} support two kinds of \code{fn}:
\itemize{
\item \emph{Transform} functions, which take an \code{xpose_data}/\code{xp_xtras} object and return one
(e.g. \code{\link{set_var_types_x}}). These are applied to each focused element in place, and
the (still-focused) \code{xpose_set} is returned so calls can keep being piped.
\item \emph{Output-generating} functions, which take an \code{xpose_data}/\code{xp_xtras} object but return
something else (e.g. a plot or table). These are applied to each focused element, and
the raw output is returned instead of an \code{xpose_set}: a single value if only one element
is focused, or a named list (by label) of outputs if several are focused.
}
}
\examples{

# Select two xpdb objects to focus on
xpdb_set \%>\% focus_xpdb(mod2,fix1)

# Add a focus
xpdb_set \%>\% focus_xpdb(mod2,fix1) \%>\% focus_xpdb(mod1, .add=TRUE)

# Remove focus
xpdb_set \%>\% focus_xpdb(mod2,fix1) \%>\% focus_xpdb()

\dontrun{
# Focus function and tidyselect
pheno_set \%>\%
  focus_xpdb(everything()) \%>\%
  # Add iOFV col and iofv type to all xpdbs in set
  focus_function(backfill_iofv) \%>\%
  # Show 1... can do all like this, too, but no need
  unfocus_xpdb() \%>\%
  select(run6) \%>\%
  {.[[1]]$xpdb} \%>\%
  list_vars()

# Quick-apply version of previous example
pheno_set \%>\%
  focus_qapply(backfill_iofv) \%>\%
  select(run6) \%>\%
  {.[[1]]$xpdb} \%>\%
  list_vars()
}

# Output-generating function applied to a single focused element:
# returns the plot itself, not an xpose_set
pheno_set \%>\%
  focus_xpdb(run6) \%>\%
  focus_function(xpose::dv_vs_ipred)

# ... or with several elements focused, a named list of plots
pheno_set \%>\%
  focus_xpdb(run6, run7) \%>\%
  focus_function(xpose::dv_vs_ipred)

}
