% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2.R
\name{get_prm_nlmixr2}
\alias{get_prm_nlmixr2}
\title{get_prm equivalent for nlmixr2 fits}
\usage{
get_prm_nlmixr2(
  xpdb,
  transform = formals(get_prm)$transform,
  show_all = formals(get_prm)$show_all,
  quiet = FALSE
)
}
\arguments{
\item{xpdb}{<\code{xp_xtras}> With nlmixr2 fit}

\item{transform}{<\code{logical}> as in \code{\link[=get_prm]{get_prm()}}}

\item{show_all}{<\code{logical}> as in \code{\link[=get_prm]{get_prm()}}}

\item{quiet}{<\code{logical}> as in \code{\link[=get_prm]{get_prm()}}}
}
\value{
a tibble with expected columns
}
\description{
This is intended to match the <\code{xpose::get_prm}> rather than the
updated \code{\link[=get_prm]{get_prm()}}.
}
