% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_plots.R
\name{grab_xpose_plot}
\alias{grab_xpose_plot}
\title{Grab processed \code{xpose_plot}}
\usage{
grab_xpose_plot(plot)
}
\arguments{
\item{plot}{<\code{xpose_plot}> or list thereof}
}
\value{
Grob or list of grobs
}
\description{
This function is very simple and unlikely to capture
every possible situation. Paginated plots are not supported.

This is helpful for working with \code{xpose} plots in \code{patchwork} or
\code{ggpubr} functions.
}
\examples{

single_plot <- xpdb_x \%>\%
eta_vs_catcov(etavar = ETA1) \%>\%
grab_xpose_plot()

listof_plots <- xpdb_x \%>\%
eta_vs_catcov(etavar = c(ETA1,ETA3)) \%>\%
grab_xpose_plot()

}
