% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xset_features.R
\name{\%p\%}
\alias{\%p\%}
\title{Binary check if LHS is parent of RHS}
\usage{
possible_parent \%p\% possible_child
}
\arguments{
\item{possible_parent}{<\code{xpose_set_item}> object suspected as parent to ...}

\item{possible_child}{... <\code{xpose_set_item}> object suspected child}
}
\value{
\verb{<logical>} \code{TRUE} if LHS is parent of RHS
}
\description{
Binary check if LHS is parent of RHS
}
\examples{

# Detect direct parent
pheno_set$run6 \%p\% pheno_set$run7

# Detect non-parentage (does not try to "flip" parentage)
pheno_set$run6 \%p\% pheno_set$run5

# Does not detect grand-parentage
pheno_set$run6 \%p\% pheno_set$run13

}
