% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xset_plots.R
\name{is_franken_xpdb}
\alias{is_franken_xpdb}
\title{Is this a model-averaged/combined ("franken") \code{xpose_data} object?}
\usage{
is_franken_xpdb(xpdb)
}
\arguments{
\item{xpdb}{<\code{xpose_data}> or <\code{xp_xtras}> object}
}
\value{
<\code{logical}>
}
\description{
Is this a model-averaged/combined ("franken") \code{xpose_data} object?
}
\keyword{internal}
