% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{is_xp_xtras}
\alias{is_xp_xtras}
\title{Basic class checker for \code{xp_xtras}}
\usage{
is_xp_xtras(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
<\code{logical}> TRUE if \code{xp_xtras} object
}
\description{
Basic class checker for \code{xp_xtras}
}
\examples{

is_xp_xtras(xpose::xpdb_ex_pk)

is_xp_xtras(xpdb_x)

}
