% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{list_dv_probs}
\alias{list_dv_probs}
\title{For a categorical DV variable, show associated probabilities}
\usage{
list_dv_probs(xpdb, .problem = NULL, .dv_var = NULL)
}
\arguments{
\item{xpdb}{<\code{xp_xtras}> object}

\item{.problem}{<\code{numeric}> Problem number to use. Uses all problems if \code{NULL}
(the default). May be omitted entirely and left to default, even when
formulas are supplied positionally in \code{...}.}

\item{.dv_var}{<\code{tidyselect}> of column having the categorical observation. Default is first-listed
\code{catdv}.}
}
\value{
<\code{tibble}> of probabilities
}
\description{
A convenient quick check for how probabilities are currently
assigned, based on \code{\link{set_dv_probs}}.
}
\examples{

pkpd_m3 \%>\%
  set_dv_probs(1, 1~LIKE, .dv_var = BLQ) \%>\%
  list_dv_probs(.dv_var=BLQ)

}
