% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stats.R
\name{logLik.xpose_data}
\alias{logLik.xpose_data}
\title{Log-likelihood, AIC and BIC for \code{xpose_data} objects}
\usage{
\method{logLik}{xpose_data}(object, .problem = NULL, .subprob = NULL, .method = NULL, ...)
}
\arguments{
\item{object}{<\code{xpose_data}> or <\code{xp_xtras}> object}

\item{.problem}{<\code{numeric}> Problem number to use. Uses the xpdb default if not provided.}

\item{.subprob}{<\code{numeric}> Subproblem number to use. Uses the xpdb default if not provided.}

\item{.method}{<\code{numeric}> Method to use. Uses the xpdb default if not provided.}

\item{...}{Not used.}
}
\value{
<\code{logLik}> object, as documented in <\code{\link[stats:logLik]{stats::logLik}}>, with \code{"df"}
(number of estimated parameters) and \code{"nobs"} (number of observations) attributes set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

NONMEM (and nlmixr2) objective function values (OFV) are \code{-2*log-likelihood},
so the log-likelihood of a model is derived as \code{-ofv/2}. The number of estimated
parameters (thetas, and unfixed omegas/sigmas) is used as the degrees of freedom.

Because <\code{\link[stats:AIC]{stats::AIC}}> and <\code{\link[stats:BIC]{stats::BIC}}> dispatch through <\code{\link[stats:logLik]{stats::logLik}}>
by default, only \code{logLik.xpose_data} needs to be defined here for \code{AIC()}/\code{BIC()}
to work as expected on \code{xpose_data}/\code{xp_xtras} objects; no \code{AIC}/\code{BIC} methods
are defined for a single model.

Not calculable for a model-averaged ("franken") \code{xpose_data} object (eg, the
output of <\code{\link{modavg_xpdb}}>), since such an object does not correspond to a
single fitted model; an error is thrown instead.
}
\examples{

logLik(xpdb_x)
AIC(xpdb_x)
BIC(xpdb_x)

}
