% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_files}
\alias{mutate_files}
\title{Mutate the file table for an xpose data object}
\usage{
mutate_files(xpdb, ...)
}
\arguments{
\item{xpdb}{<\verb{xpose data}> object}

\item{...}{Forwarded to \link[dplyr:mutate]{mutate()}}
}
\description{
Mutate the file table for an xpose data object
}
\keyword{internal}
