% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{print.xpose_set}
\alias{print.xpose_set}
\alias{print.xpose_set_item}
\alias{c.xpose_set}
\title{Handling xpose_set objects with common methods}
\usage{
\method{print}{xpose_set}(x, ...)

\method{print}{xpose_set_item}(x, ...)

\method{c}{xpose_set}(..., .relationships = NULL)
}
\description{
Handling xpose_set objects with common methods
}
