% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlmixr2_warfarin}
\alias{nlmixr2_warfarin}
\title{An \code{xp_xtra} example based on a complex nlmixr2 fit}
\format{
An object of class \code{xp_xtras} (inherits from \code{xpose_data}, \code{uneval}) of length 10.
}
\source{
\url{https://nlmixr2.org/articles/multiple-endpoints.html}
}
\usage{
nlmixr2_warfarin
}
\description{
Based on the multiple endpoint warfarin PK/PD example.
}
\seealso{
\link{xpdb_nlmixr2}
}
\keyword{datasets}
