% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_pars.R
\name{param_selector}
\alias{param_selector}
\title{Select parameter row number}
\usage{
param_selector(sel, prm_tbl)
}
\arguments{
\item{sel}{<\code{character}> Selector of any parameter}

\item{prm_tbl}{<\code{tibble}> Like the output of \code{get_prm()}}
}
\value{
<\code{integer}> of selected row number. Can be used to
get value and other elements of information from \code{prm_tbl}.
}
\description{
The selection rules are described in <\code{\link{add_prm_association}}>.
}
\keyword{internal}
