% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_base}
\alias{pheno_base}
\title{An \code{xp_xtras} example of a base model}
\format{
\subsection{\code{xp_xtras}}{

An \code{xp_xtras} object.
}
}
\source{
https://doi.org/10.1159/000457062 and nlmixr2data::pheno_sd
}
\usage{
pheno_base
}
\description{
Base model for phenobarbital in neotates.
}
\details{
This is \code{run6} in <\code{\link{pheno_set}}>
}
\keyword{datasets}
