% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_final}
\alias{pheno_final}
\title{An \code{xp_xtras} example of a final model}
\format{
\subsection{\code{xp_xtras}}{

An \code{xp_xtras} object.
}
}
\source{
https://doi.org/10.1159/000457062 and nlmixr2data::pheno_sd
}
\usage{
pheno_final
}
\description{
Final model for phenobarbital in neotates.
}
\details{
This is re-parameterized from the covariate-building work, which
in this case did not identify a relationship with Apgar score.

This is \code{run16} in <\code{\link{pheno_set}}>
}
\keyword{datasets}
