% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_saem}
\alias{pheno_saem}
\title{An \code{xp_xtras} example of a final model}
\format{
\subsection{\code{xp_xtras}}{

An \code{xp_xtras} object.
}
}
\source{
\url{https://doi.org/10.1159/000457062} and nlmixr2data::pheno_sd
}
\usage{
pheno_saem
}
\description{
Final model for phenobarbital in neonates.
}
\details{
This is the same as \code{\link{pheno_final}} but fitted with SAEM/IMP.

Not a part of <\code{\link{pheno_set}}>
}
\keyword{datasets}
