% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_set}
\alias{pheno_set}
\title{A more complex example of \code{xpose_set} object}
\format{
\subsection{\code{xpose_set}}{

An \code{xpose_set} object of length 14 with a branched lineage.
}
}
\source{
\url{https://doi.org/10.1159/000457062} and nlmixr2data::pheno_sd
}
\usage{
pheno_set
}
\description{
Model-building set for the phenobarbital in neonates PK data
used across multiple packages.
}
\details{
This is not a demonstration of high-quality model-building,
it is just a typical and simple example.
}
\keyword{datasets}
