% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkpd_m3}
\alias{pkpd_m3}
\title{An \code{xp_xtras} example of an M3 model}
\format{
\subsection{\code{xp_xtras}}{

An \code{xp_xtras} object.
}
}
\source{
https://doi.org/10.1002/psp4.13219
}
\usage{
pkpd_m3
}
\description{
A representative PK/PD model with M3 fitting applied.
}
\references{
Beal, S.L. Ways to Fit a PK Model with Some Data
Below the Quantification Limit. J Pharmacokinet
Pharmacodyn 28, 481-504 (2001).
https://doi.org/10.1023/A:1012299115260

Prybylski JP. Indirect modeling of derived outcomes:
Are minor prediction discrepancies a cause for concern?
CPT Pharmacometrics Syst Pharmacol. 2024; 00: 1-9.
doi:10.1002/psp4.13219
}
\keyword{datasets}
