% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkpd_m3_df}
\alias{pkpd_m3_df}
\title{An \code{xp_xtras} example of an M3 model (dataset)}
\format{
\subsection{\code{xp_xtras}}{

An \code{xp_xtras} object.
}
}
\source{
\url{https://doi.org/10.1002/psp4.13219}
}
\usage{
pkpd_m3_df
}
\description{
The dataset used to fit the \code{\link{pkpd_m3}} model.
}
\references{
Prybylski JP. Indirect modeling of derived outcomes:
Are minor prediction discrepancies a cause for concern?
CPT Pharmacometrics Syst Pharmacol. 2024; 00: 1-9.
doi:10.1002/psp4.13219
}
\keyword{datasets}
