% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{proc_levels}
\alias{proc_levels}
\title{Convert levels list into tibble}
\usage{
proc_levels(lvl_list)
}
\arguments{
\item{lvl_list}{}
}
\value{
<\code{tibble}> of levels
}
\description{
Consumes formula list and converts into corresponding tibble.
}
