% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{pull.xpose_set}
\alias{pull.xpose_set}
\title{Pulling method for xpose_set}
\usage{
\method{pull}{xpose_set}(.data, ...)
}
\arguments{
\item{.data}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> (passed through to <\code{\link[dplyr:pull]{pull}}>)}
}
\value{
The top-level information for a set requested.
}
\description{
Pulling method for xpose_set
}
\examples{

xpdb_set \%>\%
  pull(xpdb)

}
