% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{rename.xpose_set}
\alias{rename.xpose_set}
\title{Renaming method for xpose_set}
\usage{
\method{rename}{xpose_set}(.data, ...)
}
\arguments{
\item{.data}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> (passed indirectly to <\code{\link[dplyr:mutate]{dplyr::mutate}}>)}
}
\value{
Re-labeled set
}
\description{
Renaming method for xpose_set
}
\examples{
xpdb_set \%>\%
  rename(Mod = mod1)


}
