% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reportable_digits}
\alias{reportable_digits}
\title{Reportable digits for model fit}
\usage{
reportable_digits(xpdb, .default = 3, .problem, .subprob, .method)
}
\arguments{
\item{xpdb}{<\code{xpose_data}\link[xpose:xpose_data]{xpose::xpose_data}> object}

\item{.default}{<\code{numeric}> Default number of digits to return if not found}

\item{.problem}{<\code{numeric}> Problem number to use. Uses all problem if not provided.}

\item{.subprob}{<\code{numeric}> Subproblem number to use. Uses the xpose default if not provided.}

\item{.method}{<\code{character}> Method to use. Uses the xpose default if not provided.}
}
\value{
Number of reportable digits
}
\description{
An opinionated function where for optimization routines
that report number of significant digits (eg, FO-based), only
those number of digits are considered reportable.
}
\examples{

reportable_digits(xpdb_x)

}
