% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{select.xpose_set}
\alias{select.xpose_set}
\title{Selection method for xpose_set}
\usage{
\method{select}{xpose_set}(.data, ...)
}
\arguments{
\item{.data}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> (passed through to <\code{\link{select_subset}}>)}
}
\description{
Selection method for xpose_set
}
\examples{
xpdb_set \%>\%
  select(starts_with("fix"))

xpdb_set \%>\%
  select(mod1, fix1)

}
