% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_set.R
\name{select_subset}
\alias{select_subset}
\title{Convenience wrapper for tidyselect}
\usage{
select_subset(xpdb_s, ...)
}
\arguments{
\item{xpdb_s}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more tidyselect selectors}
}
\value{
<\code{\link{numeric}}> vector of indices for selected xpdb objects
}
\description{
This is intended for use as an internal function to select a subset of xpdb objects from an xpose_set.
}
\examples{

data("xpdb_ex_pk", package = "xpose")

# Arbitrary copies
xpdb_ex_pk2 <- xpdb_ex_pk3 <- xpdb_ex_pk4 <- xpdb_ex_pk

set <- xpose_set(mod1=xpdb_ex_pk, mod2=xpdb_ex_pk2, fix1 = xpdb_ex_pk3, fix2 = xpdb_ex_pk4)

xpose.xtras:::select_subset(mod2, xpdb_s=set)

xpose.xtras:::select_subset(dplyr::starts_with("fix"), xpdb_s=set)


}
\keyword{internal}
