% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_features.R
\name{set_base_model}
\alias{set_base_model}
\alias{get_base_model}
\alias{unset_base_model}
\title{Base model for \code{xpose_set}}
\usage{
set_base_model(xpdb_s, ...)

get_base_model(xpdb_s)

unset_base_model(xpdb_s)
}
\arguments{
\item{xpdb_s}{<\code{xpose_set}> object}

\item{...}{<<\code{\link[rlang:dyn-dots]{dynamic-dots}}> name of base model}
}
\value{
<\code{xpose_set}> object with a base model
}
\description{
Base model for \code{xpose_set}
}
\examples{

w_base <- xpdb_set \%>\%
  set_base_model(mod2)
w_base # base model listed in output

get_base_model(w_base) # base model name

unset_base_model(w_base) # base model no longer in output


}
