% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{shk_vs_contcov}
\alias{shk_vs_contcov}
\title{Shrinkage contribution versus continuous covariates}
\usage{
shk_vs_contcov(
  xpdb,
  mapping = NULL,
  shkvar = NULL,
  drop_fixed = TRUE,
  linsm = FALSE,
  type = "ps",
  title = "Shrinkage contribution versus continuous covariates | @run",
  subtitle = "Based on @nind individuals",
  caption = "@dir",
  tag = NULL,
  log = NULL,
  guide = TRUE,
  facets,
  .problem,
  quiet,
  ...
)
}
\arguments{
\item{xpdb}{<\code{xp_xtras}> or <\code{xpose_data}> object}

\item{mapping}{\code{ggplot2} style mapping}

\item{shkvar}{\code{tidyselect} for \code{shk} variables}

\item{drop_fixed}{As in \code{xpose}}

\item{linsm}{If \code{type} contains "s" should the smooth method by \code{lm}?}

\item{type}{Passed to \code{xplot_scatter}}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}

\item{tag}{Plot tag}

\item{log}{Log scale covariate value?}

\item{guide}{Add guide line?}

\item{facets}{Additional facets}

\item{.problem}{Problem number}

\item{quiet}{Silence output}

\item{...}{Any additional aesthetics.}
}
\value{
The desired plot
}
\description{
Mirrors \code{\link[=eta_vs_contcov]{eta_vs_contcov()}}, but for the per-individual shrinkage
contribution diagnostic (\code{shk} type columns, see
\code{\link[=derive_shk]{derive_shk()}}/\code{\link[=backfill_shk]{backfill_shk()}}) instead of the etas themselves.
}
\examples{
\donttest{

xpdb_x \%>\%
  backfill_shk() \%>\%
  shk_vs_contcov()
}
}
