% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2.R
\name{test_nlmixr2_has_fit}
\alias{test_nlmixr2_has_fit}
\title{Test if xpose data object has a fit object}
\usage{
test_nlmixr2_has_fit(xpdb)
}
\arguments{
\item{xpdb}{<\code{xpose_data}> object}
}
\description{
Test if xpose data object has a fit object
}
\keyword{internal}
