% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_xpdb}
\alias{test_xpdb}
\title{Logical instead of exception for xpose data check}
\usage{
test_xpdb(xpdb, check = "data")
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{check}{The `xpdb` slot to be checked, can be `data`, `files`, 
`summary`, `special`, `code`, a file name, or `FALSE`. IF `FALSE` the slot 
test will be skipped.}
}
\description{
Logical instead of exception for xpose data check
}
\keyword{internal}
