% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xset_plots.R
\name{two_set_dots}
\alias{two_set_dots}
\title{Typical processing for plots of 2 sets}
\usage{
two_set_dots(xpdb_s, ..., .inorder = FALSE, envir = parent.frame())
}
\arguments{
\item{xpdb_s}{<\code{xpose_set}> object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Passed to <[\code{select_subset}>}

\item{.inorder}{<\code{logical}> Regardless of base model or parentage, use the
two plots in order of how they are in arguments. First plot listed is treated
as base or parent.}

\item{envir}{Where to assign \code{mod1} and \code{mod2} <\code{xpose_set_item}>s}
}
\value{
Into environment specified by \code{envir}, \code{mod1} and \code{mod2}
}
\description{
Typical processing for plots of 2 sets
}
