% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{val2lvl}
\alias{val2lvl}
\title{Translate values to levels}
\usage{
val2lvl(vals, lvl_tbl = NULL)
}
\arguments{
\item{vals}{vector of values associated with levels in \code{lvl_tbl}}

\item{lvl_tbl}{tibble of levels}
}
\value{
A vector of levels corresponding to the input vector. If
\code{lvl_tbl} carries an \code{ordered} attribute set to \code{TRUE} (see
\code{\link[=set_var_levels]{set_var_levels()}}'s \code{.ordered} argument and \code{\link[=lvl_inord]{lvl_inord()}}), the
result is an ordered factor.
}
\description{
This is intended to be used as a convenience function
in plotting where levels are set for some variable.
}
