% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vismo_dtmm}
\alias{vismo_dtmm}
\title{An \code{xp_xtras} example of the discrete-time Markov model of categorical vismodegib data}
\format{
\subsection{\code{xp_xtras}}{

An \code{xp_xtras} object.
}
}
\source{
Derived from sup-0009 and sup-0010 from the reference.
}
\usage{
vismo_dtmm
}
\description{
The referenced work presents two alternative modeling approaches
for muscle spasm response to vismodegib. This is a fit of the
provided discrete-time Markov model to the 50 participant mock data.
}
\references{
Lu, T., Yang, Y., Jin, J.Y. and Kågedal, M. (2020),
Analysis of Longitudinal-Ordered Categorical Data for
Muscle Spasm Adverse Event of Vismodegib: Comparison
Between Different Pharmacometric Models. CPT
Pharmacometrics Syst. Pharmacol., 9: 96-105.
\url{https://doi.org/10.1002/psp4.12487}
}
\keyword{datasets}
