% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vismodegib}
\alias{vismodegib}
\title{A tibble of mock data used for fitting vismodegib models}
\format{
\subsection{\code{tibble}}{

An \code{tibble}.
}
}
\source{
Generated using sup-0009 and sup-0010 from the reference.
}
\usage{
vismodegib
}
\description{
The referenced work presents two alternative modeling approaches
for muscle spasm response to vismodegib. There is a mock
dataset for one person, and using the provided model a 50 participant
mock dataset could be generated.
}
\references{
Lu, T., Yang, Y., Jin, J.Y. and Kågedal, M. (2020),
Analysis of Longitudinal-Ordered Categorical Data for
Muscle Spasm Adverse Event of Vismodegib: Comparison
Between Different Pharmacometric Models. CPT
Pharmacometrics Syst. Pharmacol., 9: 96-105.
https://doi.org/10.1002/psp4.12487
}
\keyword{datasets}
