% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{where_xp}
\alias{where_xp}
\title{Alternative to where() for \code{xpose_set}}
\usage{
where_xp(fn)
}
\arguments{
\item{fn}{A function that returns \code{TRUE} or \code{FALSE} (technically, a
\emph{predicate} function). Can also be a purrr-like formula.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

For use when <\code{\link[tidyselect:where]{where}}> might be useful. This is a work in progress.

Unlikely to offer any benefits over <\code{\link{reshape_set}}> and then using typical <\code{\link[tidyselect:where]{where}}>.
}
\examples{

xpose.xtras:::select_subset(where_xp(~"fix1" \%in\% parent), xpdb_s=xpdb_set)
}
