% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_plots.R
\name{wrap_xp_ggally}
\alias{wrap_xp_ggally}
\title{Ensure consistent style with \code{GGally} functions}
\usage{
wrap_xp_ggally(fn, mapping = NULL, xp_theme, ...)
}
\arguments{
\item{fn}{<\code{character}> name of \code{GGally} function}

\item{mapping}{<\code{ggplot2::aes}> mapping}

\item{xp_theme}{theme to use}

\item{...}{<\code{any}> additional arguments to pass to \code{GGally} function}
}
\value{
\code{ggplot2} function
}
\description{
Ensure consistent style with \code{GGally} functions
}
