% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_plots.R
\name{xp4_xtra_theme}
\alias{xp4_xtra_theme}
\title{Updated version of the xpose4 theme}
\usage{
xp4_xtra_theme()
}
\value{
An \code{xpose} theme object with \code{xpose4} color palette
}
\description{
Updated version of the xpose4 theme
}
