% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{xp_var}
\alias{xp_var}
\alias{xp_var.default}
\alias{xp_var.xp_xtras}
\title{\code{xp_var} Method}
\usage{
xp_var(xpdb, .problem, col = NULL, type = NULL, silent = FALSE)

\method{xp_var}{default}(xpdb, .problem, col = NULL, type = NULL, silent = FALSE)

\method{xp_var}{xp_xtras}(xpdb, .problem, col = NULL, type = NULL, silent = FALSE)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{.problem}{The $problem number to be used.}

\item{col}{The column name to be searched in the index. Alternative to arg `type`.}

\item{type}{The type of column to searched in the index. Alternative to `col`.}

\item{silent}{Should the function be silent or return errors.}
}
\value{
A tibble of identified variables.
}
\description{
To add a small amount of functionality to <\code{\link[xpose:xp_var]{xp_var}}>,
this method was created.
}
