% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_plots.R
\name{xp_xtra_theme}
\alias{xp_xtra_theme}
\title{Extra theme defaults}
\usage{
xp_xtra_theme(base_on = NULL)
}
\arguments{
\item{base_on}{\code{xp_theme} object to extend}
}
\description{
Adds aesthetics for plot components used in this
package.
}
\details{
This package attempts to generate a consistent
theme even if users are working with a highly
cutomized \code{xp_theme}. There is are only a few
hard-coded aesthetics, and the rest are derived from
existing aesthetics in \code{base_on}, which defaults to
the default from \code{xpose}.

Only a few options are worth noting. In <\code{\link{xplot_pairs}}>
(and functions using it), the aesthetics for \code{GGally}-specific
elements like \code{barDiag} are defined as \verb{gga(element)_(aesthetic)}.
The labeller for pairs plots is also changed from the \emph{de facto} default
\code{label_both} to \code{label_value}, but any labeller can be provided as
\code{pairs_labeller}.
}
