% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xpdb_nlmixr2}
\alias{xpdb_nlmixr2}
\alias{xpdb_nlmixr2_saem}
\title{An \code{xp_xtra} example based on an nlmixr2 fit}
\format{
An object of class \code{xp_xtras} (inherits from \code{xpose_data}, \code{uneval}) of length 10.

An object of class \code{xp_xtras} (inherits from \code{xpose_data}, \code{uneval}) of length 10.
}
\source{
\url{https://nlmixr2.org/articles/running_nlmixr.html}
}
\usage{
xpdb_nlmixr2

xpdb_nlmixr2_saem
}
\description{
This is the most basic one compartment example used in
nlmixr2 example documentation. It is a fit to the popular theophylline
dataset.
}
\references{
Fidler M (2025). \emph{nlmixr2: Nonlinear Mixed Effects Models in Population PK/PD}.
doi:10.32614/CRAN.package.nlmixr2 \url{https://doi.org/10.32614/CRAN.package.nlmixr2},
R package version 3.0.2, \url{https://CRAN.R-project.org/package=nlmixr2}.

Fidler M, Wilkins J, Hooijmaijers R, Post T, Schoemaker R, Trame M, Xiong Y, Wang W
(2019). “Nonlinear Mixed-Effects Model Development and Simulation Using nlmixr and
Related R Open-Source Packages.” \emph{CPT: Pharmacometrics & Systems Pharmacology},
\emph{8}(9), 621-633. \url{https://doi.org/10.1002/psp4.12445}.

Schoemaker R, Fidler M, Laveille C, Wilkins J, Hooijmaijers R, Post T, Trame M,
Xiong Y, Wang W (2019). “Performance of the SAEM and FOCEI Algorithms in the
Open-Source, Nonlinear Mixed Effect Modeling Tool nlmixr.” \emph{CPT: Pharmacometrics &
Systems Pharmacology}, \emph{8}(12), 923-930. \url{https://doi.org/10.1002/psp4.12471}.
}
\seealso{
\link[datasets:Theoph]{Theoph}
}
\keyword{datasets}
