% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xpdb_set}
\alias{xpdb_set}
\title{An example \code{xpose_set} object}
\format{
\subsection{\code{xpose_set}}{

An \code{xpose_set} object of length 4 with a single lineage.
}
}
\source{
Assembled from the \code{xpdb_ex_pk} object in the \code{xpose} package.
}
\usage{
xpdb_set
}
\description{
A set of identical \code{xpdb} objects to demo various features of \code{xpose.xtras}.
}
\keyword{datasets}
