% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xpdb_x}
\alias{xpdb_x}
\title{An example \code{xp_xtras} object}
\format{
\subsection{\code{xp_xtras}}{

An \code{xp_xtras} object with no extra data filled.
}
}
\source{
Assembled from the \code{xpdb_ex_pk} object in the \code{xpose} package.
}
\usage{
xpdb_x
}
\description{
The <\code{\link[xpose:xpdb_ex_pk]{xpdb_ex_pk}}> object coverted to \code{xp_xtras}. For examples.
}
\keyword{datasets}
