% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_plots.R
\name{xplot_pairs}
\alias{xplot_pairs}
\title{Wrapper around ggpairs}
\usage{
xplot_pairs(
  xpdb,
  mapping = NULL,
  group = "ID",
  type = "pls",
  guide = FALSE,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  gg_theme,
  xp_theme,
  opt,
  quiet,
  ...
)
}
\value{
specified pair plot
}
\description{
Following the \code{xpose} design pattern to derive <\code{\link[GGally:ggpairs]{ggpairs}}> plots.

Established \code{xplot_} are used to generate parts of the grid.
}
\examples{
c()

}
