% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_features.R
\name{xset_lineage}
\alias{xset_lineage}
\title{Determine lineage within a set}
\usage{
xset_lineage(xpdb_s, ...)
}
\arguments{
\item{xpdb_s}{<\code{xpose_set}> object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> either labels for
models in the set from which to create lineage lists. If empty,
lineage from base model will be output; if no base, first listed
model will be used. Always used the most senior model in this list.}
}
\value{
<\code{character}> vector of \code{c('base', 'base child', 'base grandchild', ...)} or list thereof, depending on dots arguments.
}
\description{
Determine lineage within a set
}
\examples{

c()

}
