% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{check_levels}
\alias{check_levels}
\title{Verify validity of level list}
\usage{
check_levels(lvl_list, index)
}
\arguments{
\item{lvl_list}{<\code{list}> of formulas or leveler functions}

\item{index}{Index of \code{xp_xtras} object}
}
\value{
Nothing, warning or error
}
\description{
Verify validity of level list
}
