% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_set.R
\name{check_xpose_set}
\alias{check_xpose_set}
\alias{check_xpose_set_item}
\title{Check an \code{xpose_set} object}
\usage{
check_xpose_set(xpdb_s)

check_xpose_set_item(xpdb_s_i)
}
\arguments{
\item{xpdb_s}{<\code{\link{xpose_set}}> An xpose_set object}

\item{xpdb_s_i}{<\code{\link[=xpose_set]{xpose_set_item}}> An xpose_set_item object (element of an xpose_set)}
}
\value{
TRUE or error thrown
}
\description{
Check an \code{xpose_set} object
}
\examples{

check_xpose_set(xpdb_set)

check_xpose_set_item(xpdb_set$mod1)

}
