% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_features.R
\name{diff.xpose_set}
\alias{diff.xpose_set}
\title{Display deltaOFV values across \code{xpose_set}}
\usage{
\method{diff}{xpose_set}(xpdb_s)
}
\arguments{
\item{xpdb_s}{<\code{xpose_set}> object}
}
\value{
<\code{numeric}> vector of deltaOFV values
}
\description{
If no base model is provided, and if lineage is unclear,
the first model in the \code{xpose_set} is used as the base model.
}
\examples{

c()
}
