% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_plots.R
\name{eta_vs_catcov}
\alias{eta_vs_catcov}
\title{Eta categorical covariate plots (typical)}
\usage{
eta_vs_catcov(
  xpdb,
  mapping = NULL,
  etavar = NULL,
  drop_fixed = TRUE,
  orientation = "x",
  show_n = check_xpdb_x(xpdb, .warn = FALSE),
  type = "bol",
  title = "Eta versus categorical covariates | @run",
  subtitle = "Based on @nind individuals, Eta shrink: @etashk",
  caption = "@dir",
  tag = NULL,
  facets,
  .problem,
  quiet,
  ...
)
}
\arguments{
\item{xpdb}{<\verb{xp_xtras> or  <}xpose_data`> object}

\item{mapping}{\code{ggplot2} style mapping}

\item{etavar}{\code{tidyselect} for \code{eta} variables}

\item{drop_fixed}{As in \code{xpose}}

\item{orientation}{Passed to \code{xplot_boxplot}}

\item{show_n}{Add "N=" to plot}

\item{type}{Passed to \code{xplot_boxplot}}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}

\item{tag}{Plot tag}

\item{facets}{Additional facets}

\item{.problem}{Problem number}

\item{quiet}{Silence output}

\item{...}{}
}
\description{
Eta categorical covariate plots (typical)
}
\examples{

eta_vs_contcov(xpdb_x)

}
