% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_prob_subprob_method}
\alias{fill_prob_subprob_method}
\title{Place .problem, .subprob and .method into environment consistently}
\usage{
fill_prob_subprob_method(
  xpdb,
  .problem,
  .subprob,
  .method,
  envir = parent.frame()
)
}
\arguments{
\item{xpdb}{<\code{xpose_data}> or related object}

\item{.problem}{\code{NULL} or missing}

\item{.subprob}{\code{NULL} or missing}

\item{.method}{\code{NULL} or missing}

\item{envir}{<\code{environment}> in which to assign the problem info.}
}
\description{
Since this is a common need, it is being functionalized
to ensure consistency.
}
\keyword{Internal}
