% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R
\name{focus_xpdb}
\alias{focus_xpdb}
\alias{unfocus_xpdb}
\alias{focused_xpdbs}
\alias{focus_function}
\title{Focus on an xpdb object in an xpose_set}
\usage{
focus_xpdb(xpdb_s, ..., .add = FALSE)

unfocus_xpdb(xpdb_s)

focused_xpdbs(xpdb_s)

focus_function(xpdb_s, fn, ...)
}
\arguments{
\item{xpdb_s}{<\code{\link{xpose_set}}> An xpose_set object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more xpdb objects to focus on}

\item{.add}{<\code{\link{logical}}> Should the focus be added to the existing focus? (default: \code{FALSE})}
}
\value{
An \code{xpose_set} object with the focused xpdb object(s)
}
\description{
For piping, set is passed, but with S3 method transformations are applied to the focused \code{xpdb} object.
}
\examples{

# Select two xpdb objects to focus on
xpdb_set \%>\% focus_xpdb(mod2,fix1)

# Add a focus
xpdb_set \%>\% focus_xpdb(mod2,fix1) \%>\% focus_xpdb(mod1, .add=TRUE)

# Remove focus
xpdb_set \%>\% focus_xpdb(mod2,fix1) \%>\% focus_xpdb()

}
