% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{as_leveler}
\alias{as_leveler}
\alias{is_leveler}
\alias{lvl_bin}
\alias{lvl_sex}
\alias{lvl_inord}
\title{Level-defining helper functions}
\usage{
as_leveler(x, .start_index = 1)

is_leveler(x)

lvl_bin(x = c("No", "Yes"), .start_index = 0)

lvl_sex()

lvl_inord(x, .start_index = 1)
}
\arguments{
\item{x}{<\code{character}> vector of levels}

\item{.start_index}{<\code{numeric}> starting index for levels}
}
\value{
Special character vector suitable to be used as leveler
}
\description{
Level-defining helper functions
}
\examples{

set_var_levels(xpdb_x,
  SEX = lvl_sex(),
  MED1 = lvl_bin(),
  MED2 = lvl_inord(c("n","y"), .start_index = 0)
  )

}
