% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_xtras.R
\name{list_vars}
\alias{list_vars}
\alias{list_vars.default}
\alias{list_vars.xp_xtras}
\title{Updates to \code{list_vars}}
\usage{
list_vars(xpdb, .problem = NULL, ...)

\method{list_vars}{default}(xpdb, .problem = NULL, ...)

\method{list_vars}{xp_xtras}(xpdb, .problem = NULL, ...)
}
\arguments{
\item{xpdb}{<\code{xpose_data}> or <\code{xp_xtras}> object}

\item{.problem}{<\code{numeric}> Problem number to use. Uses the all problems if \code{NULL}}

\item{...}{Should be blank.}
}
\value{
<\code{tibble}> of variables
}
\description{
To accommodate changes made in \code{xpose.xtras},
<\code{\link[xpose:list_vars]{list_vars}}> needed some minimal updates.
}
\examples{

list_vars(xpose::xpdb_ex_pk)

list_vars(xpdb_x)

}
