% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_set.R, R/xp_xtras.R
\name{print.xpose_set}
\alias{print.xpose_set}
\alias{print.xpose_set_item}
\alias{c.xpose_set}
\alias{print.xp_xtras}
\title{Handling xpose_set objects with common methods}
\usage{
\method{print}{xpose_set}(x, ..., n = 5)

\method{print}{xpose_set_item}(x, ...)

\method{c}{xpose_set}(..., .relationships = NULL)

\method{print}{xp_xtras}(x, ...)
}
\description{
Handling xpose_set objects with common methods
}
