% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2.R
\name{nlmixr2_as_xtra}
\alias{nlmixr2_as_xtra}
\title{Convenience function for ingesting an nlmixr2 model to xpose and xpose.xtras}
\usage{
nlmixr2_as_xtra(obj, ...)
}
\arguments{
\item{obj}{nlmixr2 fit object}

\item{...}{Passed to \link[xpose.nlmixr2:xpose_data_nlmixr2]{xpose_data_nlmixr2}}
}
\value{
An <\code{xp_xtra}> object with fit attached
}
\description{
A wrapper that executes the pipeline \verb{fit |> xpose_data_nlmixr2 |> attach_nlmixr2 |> as_xp_xtras}
}
\seealso{
\code{\link[=attach_nlmixr2]{attach_nlmixr2()}}
}
