% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_set.R
\name{reshape_set}
\alias{reshape_set}
\alias{unreshape_set}
\title{Convert xpose_set to a nested list.}
\usage{
reshape_set(x)

unreshape_set(y)
}
\arguments{
\item{x}{<\code{\link{xpose_set}}> An xpose_set object}

\item{y}{<\code{\link[tibble:tibble]{tibble}}> A nested table from an xpose_set}
}
\value{
<\code{\link[tibble:tibble]{tibble}}> Nested list, or <\code{\link{xpose_set}}>
}
\description{
This amounts to a convenience function for tidy manupulations.
}
\examples{

data("xpdb_ex_pk", package = "xpose")

# Arbitrary copies
xpdb_ex_pk2 <- xpdb_ex_pk3 <- xpdb_ex_pk4 <- xpdb_ex_pk

set <- xpose_set(mod1=xpdb_ex_pk, mod2=xpdb_ex_pk2, fix1 = xpdb_ex_pk3, fix2 = xpdb_ex_pk4)

rset <- reshape_set(set)
# Properties (exposed and top-level) can be seen. xpdb objects are nested in the xpdb column.
rset \%>\% dplyr::select(-xpdb) \%>\% dplyr::glimpse()

unreshape_set(rset)

}
