% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_labs.R
\name{set_default_labs}
\alias{set_default_labs}
\title{Set default plot label overrides on an \code{xp_xtras} object}
\usage{
set_default_labs(xpdb, ...)
}
\arguments{
\item{xpdb}{<\code{\link[xpose:xpose_data]{xpose_data}}> or <\code{xp_xtras}> object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more of
\code{title}, \code{subtitle}, \code{caption}, \code{tag}, given as character strings}
}
\value{
\code{xp_xtras} object
}
\description{
Stores \code{title}/\code{subtitle}/\code{caption}/\code{tag} templates on \code{xpdb} that
\code{\link[=apply_default_labs]{apply_default_labs()}} will use to fill in (or overwrite) labels on any
plot built from this \code{xpdb}, when \code{xpdb} is passed to it. Values may
contain the same \verb{@keyword} placeholders understood by
\code{\link[xpose:parse_title]{xpose::parse_title()}} (e.g. \code{"@nind"}, \code{"@nobs"}, \code{"@runno"}), since
they are resolved the same way.
}
\examples{
xpdb_x <- set_default_labs(xpdb_x, caption = "@nobs observations in @nind individuals")
p <- xpose::dv_vs_ipred(xpdb_x)
apply_default_labs(p, xpdb = xpdb_x)
}
\seealso{
\code{\link[=set_xtras_options]{set_xtras_options()}} for the session-option equivalent
(\code{xpose.xtras.default_labs}), and \code{\link[=get_xtras_option]{get_xtras_option()}} to check which
one is currently dominant.
}
