% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra_watermark.R
\name{set_default_watermark}
\alias{set_default_watermark}
\title{Set default watermark options on an \code{xp_xtras} object}
\usage{
set_default_watermark(xpdb, ...)
}
\arguments{
\item{xpdb}{<\code{\link[xpose:xpose_data]{xpose_data}}> or <\code{xp_xtras}> object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more of
\code{label}, \code{colour}, \code{alpha}, \code{size}, \code{angle}, \code{fontface}}
}
\value{
\code{xp_xtras} object
}
\description{
Stores default \code{\link[=add_watermark]{add_watermark()}} arguments on \code{xpdb} that
\code{add_watermark()} will use for any of \code{label}/\code{colour}/\code{alpha}/\code{size}/
\code{angle}/\code{fontface} not otherwise supplied, when \code{xpdb} is passed to it.
}
\examples{
xpdb_x <- set_default_watermark(xpdb_x, label = "PRELIMINARY", colour = "red")
p <- xpose::dv_vs_ipred(xpdb_x)
add_watermark(p, xpdb = xpdb_x)
}
