% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_option}
\alias{set_option}
\title{Set an \code{xpose} option}
\usage{
set_option(xpdb, ...)
}
\arguments{
\item{xpdb}{<\code{xpose_data}\link[xpose:xpose_data]{xpose::xpose_data}> object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Arguments in
the form of \code{option = value}}
}
\value{
\code{xp_xtras} object
}
\description{
Set an \code{xpose} option
}
\examples{

xpdb_x <- set_option(xpdb_x, quiet = TRUE)

}
