% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_set.R
\name{where_xp}
\alias{where_xp}
\title{Alternative to where() for \code{xpose_set}}
\usage{
where_xp(fn)
}
\arguments{
\item{fn}{A function that returns \code{TRUE} or \code{FALSE} (technically, a
\emph{predicate} function). Can also be a purrr-like formula.}
}
\description{
For use when <\link[tidyselect:where]{tidyselect::where}> might be useful.
}
\examples{

xpose.xtras:::select_subset(where.xp(~"fix1" \%in\% parent), xpdb_s=xpdb_set)
}
