% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_boxplot.R
\name{xplot_boxplot}
\alias{xplot_boxplot}
\title{Default xpose boxplot function}
\usage{
xplot_boxplot(
  xpdb,
  mapping = NULL,
  type = "bp",
  xscale = "discrete",
  yscale = "continuous",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  plot_name = "boxplot",
  intercept = NULL,
  orientation = c("vertical", "horizontal"),
  gg_theme,
  xp_theme,
  opt,
  quiet,
  ...
)
}
\arguments{
\item{type}{<\code{character}> vector of elements to include in the plot. See \link{Details}.}

\item{orientation}{<\code{character}> Direction for the boxplot spread;
\code{xscale} and \code{yscale} will be reversed if \code{"horizontal"} is used.}
}
\description{
Manually generate boxplots from an xpdb object.
}
\details{
For type-based customization of plots:
\itemize{
\item \code{b} box-whisker (using default quantiles)
\item \code{p} points (from geom_dotplot)
\item \code{v} violin (from geom_violin)
\item \code{o} outliers (show outliers)
}
}
