% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_pairs.R
\name{xplot_pairs}
\alias{xplot_pairs}
\title{Wrapper around ggpairs}
\usage{
xplot_pairs(
  xpdb,
  mapping = NULL,
  vars = NULL,
  cont_opts = list(group = "ID", guide = FALSE, type = "ps"),
  dist_opts = list(guide = FALSE, type = "hr"),
  cat_opts = list(type = "bo", log = NULL),
  contcont_opts = list(other_fun = NULL, stars = FALSE, digits = reportable_digits(xpdb),
    title = "Pearson Corr"),
  catcont_opts = list(other_fun = NULL, stars = FALSE, digits = reportable_digits(xpdb),
    title = "Spearman rho"),
  catcat_opts = list(use_rho = TRUE),
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  plot_name = "pairs",
  gg_theme,
  xp_theme,
  opt,
  quiet,
  ...
)
}
\value{
specified pair plot
}
\description{
Following the \code{xpose} design pattern to derive <\code{\link[GGally:ggpairs]{ggpairs}}> plots.

Established \code{xplot_} are used to generate parts of the grid.
}
\examples{
c()

}
