% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.xtras-package.R
\docType{package}
\name{xpose.xtras-package}
\alias{xpose.xtras}
\alias{xpose.xtras-package}
\title{xpose.xtras: Extra Functionality for the Xpose Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Adding some at-present missing functionality, or functions unlikely to be added to the base xpose package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jprybylski.github.io/xpose.xtras/}
}

}
\author{
\strong{Maintainer}: John Prybylski \email{john.prybylski@pfizer.com} (\href{https://orcid.org/0000-0001-5802-0539}{ORCID})

}
\keyword{internal}
